🚚 zb: Move vsock socket type impls to their own module
